+++ gsh/StaticSum.hpp
@@ -6,15 +6,15 @@ template<class T, class F = gsh::Plus, class I = gsh::Negate, class Alloc = std:
   [[no_unique_address]] F func;
   [[no_unique_address]] I inv;
   constexpr void build() {
-    for(i64 i = 0; i + 1 < prefix_sum.size(); ++i) { prefix_sum[i + 1] = std::invoke(func, prefix_sum[i], prefix_sum[i + 1]); }
+    for(u32 i = 0; i + 1 < prefix_sum.size(); ++i) { prefix_sum[i + 1] = std::invoke(func, prefix_sum[i], prefix_sum[i + 1]); }
   }
 public:
   using reference = T&;
   using const_reference = const T&;
   using iterator = T*;
   using const_iterator = const T*;
-  using size_type = i64;
-  using difference_type = i64;
+  using size_type = u32;
+  using difference_type = i32;
   using value_type = T;
   using allocator_type = Alloc;
   constexpr StaticSum() : func(), inv() {}
@@ -30,14 +30,14 @@ public:
     prefix_sum.assign(std::ranges::begin(r), std::ranges::end(r));
     build();
   }
-  constexpr T sum(i64 r) const {
+  constexpr T sum(u32 r) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(r == 0) [[unlikely]]
         return T{};
     }
     return prefix_sum[r - 1];
   }
-  constexpr T sum(i64 l, i64 r) const {
+  constexpr T sum(u32 l, u32 r) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(l >= r) [[unlikely]]
         return T{};
@@ -54,7 +54,7 @@ template<class R> concept Range3D = std::ranges::forward_range<R> && Range2D<std
 } // namespace internal
 template<class T, class F = gsh::Plus, class I = gsh::Negate, class Alloc = std::allocator<T>> class StaticSum2D {
   Vec<T, Alloc> prefix_sum;
-  i64 row_size = 0;
+  u32 row_size = 0;
   [[no_unique_address]] F func;
   [[no_unique_address]] I inv;
 public:
@@ -62,8 +62,8 @@ public:
   using const_reference = const T&;
   using iterator = T*;
   using const_iterator = const T*;
-  using size_type = i64;
-  using difference_type = i64;
+  using size_type = u32;
+  using difference_type = i32;
   using value_type = T;
   using allocator_type = Alloc;
   constexpr StaticSum2D() : func(), inv() {}
@@ -78,42 +78,42 @@ public:
       row_size = 0;
       return;
     }
-    i64 h = static_cast<i64>(std::ranges::distance(itr, sent));
-    i64 w = static_cast<i64>(std::ranges::size(*itr));
+    u32 h = std::ranges::distance(itr, sent);
+    u32 w = std::ranges::size(*itr);
     prefix_sum.assign(w * h);
     row_size = w;
-    i64 idx = 0;
+    u32 idx = 0;
     while(itr != sent) {
 #ifndef NDEBUG
-      i64 sz = static_cast<i64>(std::ranges::size(*itr));
+      u32 sz = std::ranges::size(*itr);
       if(sz != row_size) [[unlikely]]
         throw Exception("gsh::StaticSum2D::set / The sizes of the rows are not the same.");
 #endif
       for(auto&& val : *itr) { prefix_sum[idx++] = std::forward<decltype(val)>(val); }
       ++itr;
     }
-    for(i64 i = 0; i < h; ++i) {
-      for(i64 j = 0; j + 1 < w; ++j) {
-        i64 idx = i * w + j;
+    for(u32 i = 0; i < h; ++i) {
+      for(u32 j = 0; j + 1 < w; ++j) {
+        u32 idx = i * w + j;
         prefix_sum[idx + 1] = std::invoke(func, prefix_sum[idx], prefix_sum[idx + 1]);
       }
     }
-    for(i64 i = 0; i + 1 < h; ++i) {
-      for(i64 j = 0; j < w; ++j) {
-        i64 idx = i * w + j;
-        i64 next_idx = (i + 1) * w + j;
+    for(u32 i = 0; i + 1 < h; ++i) {
+      for(u32 j = 0; j < w; ++j) {
+        u32 idx = i * w + j;
+        u32 next_idx = (i + 1) * w + j;
         prefix_sum[next_idx] = std::invoke(func, prefix_sum[idx], prefix_sum[next_idx]);
       }
     }
   }
-  constexpr T sum(i64 x, i64 y) const {
+  constexpr T sum(u32 x, u32 y) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(x == 0 || y == 0) [[unlikely]]
         return T{};
     }
     return prefix_sum[(x - 1) * row_size + (y - 1)];
   }
-  constexpr T sum(i64 lx, i64 ly, i64 rx, i64 ry) const {
+  constexpr T sum(u32 lx, u32 ly, u32 rx, u32 ry) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(lx >= rx || ly >= ry) [[unlikely]]
         return T{};
@@ -128,8 +128,8 @@ public:
 template<internal::Range2D R, class F = gsh::Plus, class I = gsh::Negate, class Alloc = std::allocator<std::ranges::range_value_t<std::ranges::range_value_t<R>>>> StaticSum2D(R&&, F = F(), I = I(), Alloc = Alloc()) -> StaticSum2D<std::ranges::range_value_t<std::ranges::range_value_t<R>>, F, I, Alloc>;
 template<class T, class F = gsh::Plus, class I = gsh::Negate, class Alloc = std::allocator<T>> class StaticSum3D {
   Vec<T, Alloc> prefix_sum;
-  i64 y_size = 0;
-  i64 z_size = 0;
+  u32 y_size = 0;
+  u32 z_size = 0;
   [[no_unique_address]] F func;
   [[no_unique_address]] I inv;
 public:
@@ -137,8 +137,8 @@ public:
   using const_reference = const T&;
   using iterator = T*;
   using const_iterator = const T*;
-  using size_type = i64;
-  using difference_type = i64;
+  using size_type = u32;
+  using difference_type = i32;
   using value_type = T;
   using allocator_type = Alloc;
   constexpr StaticSum3D() : func(), inv() {}
@@ -154,16 +154,16 @@ public:
       z_size = 0;
       return;
     }
-    i64 h = static_cast<i64>(std::ranges::distance(itr, sent));
-    i64 w = static_cast<i64>(std::ranges::size(*itr));
-    i64 d = static_cast<i64>(std::ranges::size(*std::ranges::begin(*itr)));
+    u32 h = std::ranges::distance(itr, sent);
+    u32 w = std::ranges::size(*itr);
+    u32 d = std::ranges::size(*std::ranges::begin(*itr));
     prefix_sum.assign(h * w * d);
     y_size = w;
     z_size = d;
-    i64 idx = 0;
+    u32 idx = 0;
     while(itr != sent) {
 #ifndef NDEBUG
-      i64 sz_w = static_cast<i64>(std::ranges::size(*itr));
+      u32 sz_w = std::ranges::size(*itr);
       if(sz_w != y_size) [[unlikely]]
         throw Exception("gsh::StaticSum3D::set / The sizes of the rows are not the same.");
 #endif
@@ -171,7 +171,7 @@ public:
       auto sent2 = std::ranges::end(*itr);
       while(itr2 != sent2) {
 #ifndef NDEBUG
-        i64 sz_d = static_cast<i64>(std::ranges::size(*itr2));
+        u32 sz_d = std::ranges::size(*itr2);
         if(sz_d != z_size) [[unlikely]]
           throw Exception("gsh::StaticSum3D::set / The sizes of the depths are not the same.");
 #endif
@@ -180,41 +180,41 @@ public:
       }
       ++itr;
     }
-    for(i64 i = 0; i < h; ++i) {
-      for(i64 j = 0; j < w; ++j) {
-        for(i64 k = 0; k + 1 < d; ++k) {
-          i64 idx = (i * w + j) * d + k;
+    for(u32 i = 0; i < h; ++i) {
+      for(u32 j = 0; j < w; ++j) {
+        for(u32 k = 0; k + 1 < d; ++k) {
+          u32 idx = (i * w + j) * d + k;
           prefix_sum[idx + 1] = std::invoke(func, prefix_sum[idx], prefix_sum[idx + 1]);
         }
       }
     }
-    for(i64 i = 0; i < h; ++i) {
-      for(i64 j = 0; j + 1 < w; ++j) {
-        for(i64 k = 0; k < d; ++k) {
-          i64 idx = (i * w + j) * d + k;
-          i64 next_idx = (i * w + (j + 1)) * d + k;
+    for(u32 i = 0; i < h; ++i) {
+      for(u32 j = 0; j + 1 < w; ++j) {
+        for(u32 k = 0; k < d; ++k) {
+          u32 idx = (i * w + j) * d + k;
+          u32 next_idx = (i * w + (j + 1)) * d + k;
           prefix_sum[next_idx] = std::invoke(func, prefix_sum[idx], prefix_sum[next_idx]);
         }
       }
     }
-    for(i64 i = 0; i + 1 < h; ++i) {
-      for(i64 j = 0; j < w; ++j) {
-        for(i64 k = 0; k < d; ++k) {
-          i64 idx = (i * w + j) * d + k;
-          i64 next_idx = ((i + 1) * w + j) * d + k;
+    for(u32 i = 0; i + 1 < h; ++i) {
+      for(u32 j = 0; j < w; ++j) {
+        for(u32 k = 0; k < d; ++k) {
+          u32 idx = (i * w + j) * d + k;
+          u32 next_idx = ((i + 1) * w + j) * d + k;
           prefix_sum[next_idx] = std::invoke(func, prefix_sum[idx], prefix_sum[next_idx]);
         }
       }
     }
   }
-  constexpr T sum(i64 x, i64 y, i64 z) const {
+  constexpr T sum(u32 x, u32 y, u32 z) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(x == 0 || y == 0 || z == 0) [[unlikely]]
         return T{};
     }
     return prefix_sum[((x - 1) * y_size + (y - 1)) * z_size + (z - 1)];
   }
-  constexpr T sum(i64 lx, i64 ly, i64 lz, i64 rx, i64 ry, i64 rz) const {
+  constexpr T sum(u32 lx, u32 ly, u32 lz, u32 rx, u32 ry, u32 rz) const {
     if constexpr(std::is_default_constructible_v<T>) {
       if(lx >= rx || ly >= ry || lz >= rz) [[unlikely]]
         return T{};
